+++ chaincfg/params_test.go
@@ -6,6 +6,17 @@ package chaincfg
 
 import "testing"
 
+// TestInvalidHashStr ensures the newShaHashFromStr function panics when used to
+// with an invalid hash string.
+func TestInvalidHashStr(t *testing.T) {
+	defer func() {
+		if r := recover(); r == nil {
+			t.Errorf("Expected panic for invalid hash, got nil")
+		}
+	}()
+	newShaHashFromStr("banana")
+}
+
 // TestMustRegisterPanic ensures the mustRegister function panics when used to
 // register an invalid network.
 func TestMustRegisterPanic(t *testing.T) {
